Added a Android version for the RAG.
This doesn't use directly the model, but rather a dict of words, and compute the cosine similarity by hand.
In later versions, another method might give better results.

(Currently, the files required for the RAG aren't included, as they will blow the size of the repo, but it's still possible to compile them by hand. Also, they won't be useful later, as the RAG will have real useful documents)